pyproject.toml
[project]
name = "crypto-data-agent"
version = "0.1.0"
description = "Real-time Gemini-powered crypto assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai-agents>=0.0.19",
  "python-dotenv>=1.0.1",
  "chainlit>=2.5.5"
]
